refactor(ProjectItemTpl): port ts
